Proyectos/Solution IPS/backend/requirements.txt
Flask==2.0.2
PyMySQL==1.0.2
Flask-Cors==3.0.10
python-dotenv==0.19.2